feat: move more types to commons (batch 20 continued)
Additional moves:
- MultiErrorToastMsg, UserBasedErrorMessage (toast multiline)
- ToastManager (toast orchestration)
- IZapField (zap split interface)
- ChannelFeedState (feed state sealed class)
- AdditiveComplexFeedFilter (feed filter)
- BookmarkType (enum)

Part of the KMP iOS migration tracked in #2238.
