+++ amethyst/src/main/java/com/vitorpamplona/amethyst/ui/components/toasts/ToastManager.kt
@@ -20,88 +20,5 @@
  */
 package com.vitorpamplona.amethyst.ui.components.toasts
 
-import androidx.compose.runtime.Stable
-import com.vitorpamplona.amethyst.model.User
-import com.vitorpamplona.amethyst.ui.components.toasts.multiline.MultiErrorToastMsg
-import com.vitorpamplona.amethyst.ui.components.toasts.multiline.UserBasedErrorMessage
-import kotlinx.coroutines.flow.MutableStateFlow
-
-@Stable
-class ToastManager {
-    val toasts = MutableStateFlow<ToastMsg?>(null)
-
-    fun clearToasts() {
-        toasts.tryEmit(null)
-    }
-
-    fun toast(
-        title: String,
-        message: String,
-    ) {
-        toasts.tryEmit(StringToastMsg(title, message))
-    }
-
-    fun toast(
-        title: String,
-        message: String,
-        action: () -> Unit,
-    ) {
-        toasts.tryEmit(ActionableStringToastMsg(title, message, action))
-    }
-
-    fun toast(
-        titleResId: Int,
-        resourceId: Int,
-    ) {
-        toasts.tryEmit(ResourceToastMsg(titleResId, resourceId))
-    }
-
-    fun toast(
-        titleResId: Int,
-        message: String?,
-        throwable: Throwable,
-    ) {
-        toasts.tryEmit(ThrowableToastMsg(titleResId, message, throwable))
-    }
-
-    fun toast(
-        titleResId: Int,
-        description: Int,
-        throwable: Throwable,
-    ) {
-        toasts.tryEmit(ThrowableToastMsg2(titleResId, description, throwable))
-    }
-
-    fun toast(
-        titleResId: Int,
-        resourceId: Int,
-        vararg params: String,
-    ) {
-        toasts.tryEmit(ResourceToastMsg(titleResId, resourceId, params))
-    }
-
-    fun toast(
-        titleResId: Int,
-        message: String,
-        user: User?,
-    ) {
-        val current = toasts.value
-        if (current is MultiErrorToastMsg && current.titleResId == titleResId) {
-            current.add(message, user)
-        } else {
-            toasts.tryEmit(MultiErrorToastMsg(titleResId).also { it.add(message, user) })
-        }
-    }
-
-    fun toast(
-        titleResId: Int,
-        data: UserBasedErrorMessage,
-    ) {
-        val current = toasts.value
-        if (current is MultiErrorToastMsg && current.titleResId == titleResId) {
-            current.add(data)
-        } else {
-            toasts.tryEmit(MultiErrorToastMsg(titleResId).also { it.add(data) })
-        }
-    }
-}
+// Re-export from commons for backwards compatibility
+typealias ToastManager = com.vitorpamplona.amethyst.commons.ui.components.toasts.ToastManager
